fix(syftbox/app/install.py): Change timeout approach to check if a repository is valid/public
Right now, we were deciding if a repository was valid by having a timeout (due to the requirement of auth in case they're private/doesn't exist). But this created flaky tests for valid prs if they take more than 2 secs to be installed. To fix that, now, we're considering every repository that requires authentication as a invalid repo